refactor: merge only specific properties
