+++ bin/cli.js
@@ -62,12 +62,12 @@ program.version(pkg.version)
 
 program.parse(process.argv)
 
-const { configFileName, configFilePath, packageFile } = program
+const { configFileName, configFilePath, packageFile, recursive } = program
 
 // load .ncurc
 // Do not load when global option is set
-// Do not load when tests are running (an be overridden if configFilePath is set explicitly)
-const rcResult = !program.global && (!process.env.NCU_TESTS || configFilePath)
+// Do not load when tests are running (an be overridden if configFilePath is set explicitly, or --recursive option specified)
+const rcResult = !program.global && (!process.env.NCU_TESTS || configFilePath || recursive)
   ? ncu.getNcurc({ configFileName, configFilePath, packageFile })
   : null
