+++ lib/index.d.ts
@@ -207,7 +207,6 @@ declare namespace ncu {
 
   function run(options?: RunOptions): Promise<RunResults>
 
-  function mergeOptions(options1: RunOptions | null, options1: RunOptions | null): RunOptions
 }
 
 export = ncu
