+++ lib/index.js
@@ -22,6 +22,7 @@ const packageManagers = require('./package-managers')
 const { print, printJson, printUpgrades, printIgnoredUpdates } = require('./logging')
 const { deepPatternPrefix, doctorHelpText } = require('./constants')
 const cliOptions = require('./cli-options')
+const mergeOptions = require('./merge-options')
 
 // maps package managers to package file names
 const packageFileNames = {
@@ -44,25 +45,6 @@ function programError(options, message) {
   else {
     throw new Error(message)
   }
-}
-
-function mergeArrays(arr1, arr2) {
-  return [...new Set([...arr1 || [], ...arr2 || []])]
-}
-
-/**
- * Merge options by a standard way, but if some properties both are arrays, then merge them also.
- */
-function mergeOptions(options1, options2) {
-  options1 = options1 || {}
-  options2 = options2 || {}
-  const result = { ...options1, ...options2 }
-  Object.keys(result).forEach(key => {
-    if (Array.isArray(options1[key]) && Array.isArray(options2[key])) {
-      result[key] = mergeArrays(options1[key], options2[key])
-    }
-  })
-  return result
 }
 
 /**
@@ -529,8 +511,8 @@ async function run(options = {}) {
         const rcResult = getNcurc({ packageFile })
         let rcConfig = rcResult && rcResult.config ? rcResult.config : {}
         if (options.recursive) {
-          // inherit parent config options
-          rcConfig = mergeOptions(options, rcConfig)
+          // Merge config options
+          rcConfig = mergeOptions(options, rcConfig, { keys: Object.keys(rcConfig) })
         }
         const pkgOptions = {
           ...options,
@@ -611,4 +593,4 @@ function getNcurc({ configFileName, configFilePath, packageFile } = {}) {
   return result ? { ...result, args } : null
 }
 
-module.exports = { run, getNcurc, getPeerDependencies, mergeOptions, ...vm }
+module.exports = { run, getNcurc, getPeerDependencies, ...vm }
